countyfrequency turns into a tinyint new photos from foster city

+++ birdwalker.sql
@@ -31,7 +31,7 @@ drop table if exists countyfrequency;
 
 CREATE TABLE countyfrequency (
   CommonName varchar(255),
-  Frequency bigint(2),
+  Frequency tinyint(2),
   SpeciesID bigint(20)
 ) TYPE=MyISAM;
